Move ingestion and evidence modules
