+++ control-plane/aegisops_control_plane/ingestion/__init__.py
@@ -1,1 +1,1 @@
-"""Package scaffold only for future ingestion lifecycle modules."""
+"""Ingestion lifecycle, case workflow, and evidence linkage modules."""
