+++ scripts/verify-phase-52-5-2-import-compatibility.sh
@@ -85,6 +85,26 @@ required_imports=(
   "aegisops_control_plane.api.http_protected_surface:authenticate_protected_write"
   "aegisops_control_plane.http_runtime_surface:runtime_read_response"
   "aegisops_control_plane.api.http_runtime_surface:runtime_read_response"
+  "aegisops_control_plane.detection_lifecycle:DetectionIntakeService"
+  "aegisops_control_plane.ingestion.detection_lifecycle:DetectionIntakeService"
+  "aegisops_control_plane.detection_lifecycle_helpers:DetectionLifecycleTransitionHelper"
+  "aegisops_control_plane.ingestion.detection_lifecycle_helpers:DetectionLifecycleTransitionHelper"
+  "aegisops_control_plane.detection_native_context:NativeDetectionContextAttacher"
+  "aegisops_control_plane.ingestion.detection_native_context:NativeDetectionContextAttacher"
+  "aegisops_control_plane.case_workflow:CaseWorkflowService"
+  "aegisops_control_plane.ingestion.case_workflow:CaseWorkflowService"
+  "aegisops_control_plane.evidence_linkage:EvidenceLinkageService"
+  "aegisops_control_plane.ingestion.evidence_linkage:EvidenceLinkageService"
+  "aegisops_control_plane.external_evidence_boundary:ExternalEvidenceBoundary"
+  "aegisops_control_plane.evidence.external_evidence_boundary:ExternalEvidenceBoundary"
+  "aegisops_control_plane.external_evidence_facade:ExternalEvidenceFacade"
+  "aegisops_control_plane.evidence.external_evidence_facade:ExternalEvidenceFacade"
+  "aegisops_control_plane.external_evidence_misp:MispExternalEvidenceHelper"
+  "aegisops_control_plane.evidence.external_evidence_misp:MispExternalEvidenceHelper"
+  "aegisops_control_plane.external_evidence_osquery:OsqueryExternalEvidenceHelper"
+  "aegisops_control_plane.evidence.external_evidence_osquery:OsqueryExternalEvidenceHelper"
+  "aegisops_control_plane.external_evidence_endpoint:EndpointExternalEvidenceHelper"
+  "aegisops_control_plane.evidence.external_evidence_endpoint:EndpointExternalEvidenceHelper"
 )
 
 if [[ ! -f "${doc_path}" ]]; then
